moved failure class into separated classes
